requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
supabase==2.3.4
python-dotenv==1.0.0
httpx==0.25.2